#67 Turned PresentationSelection.jsx into PresentationSelection.tsx for consistency
